updating DisableWindowsDefender

+++ 00_Windows/DisableWindowsDefender.ps1
@@ -1,42 +1,40 @@
 # Disable Windows Defender
-ForceRegKey("HKLM:\Software\Policies\Microsoft\Windows Defender")
-Set-ItemProperty -Path "HKLM:\Software\Policies\Microsoft\Windows Defender" -Name "DisableAntiSpyware" -Type DWord -Value 1
+if (Test-Path "HKLM:\Software\Policies\Microsoft\Windows Defender") {
+    Set-ItemProperty -Path "HKLM:\Software\Policies\Microsoft\Windows Defender" -Name "DisableAntiSpyware" -Type DWord -Value 1
 
+    $tasks = @(
+        "\Microsoft\Windows\Windows Defender\Windows Defender Cache Maintenance"
+        "\Microsoft\Windows\Windows Defender\Windows Defender Cleanup"
+        "\Microsoft\Windows\Windows Defender\Windows Defender Scheduled Scan"
+        "\Microsoft\Windows\Windows Defender\Windows Defender Verification"
+    )
 
+    foreach ($task in $tasks) {
+        $parts = $task.split('\')
+        $name = $parts[-1]
+        $path = $parts[0..($parts.length-2)] -join '\'
+        Remove-ScheduledTask -TaskName "$name" -TaskPath "$path"
+    }
 
-$tasks = @(
-    "\Microsoft\Windows\Windows Defender\Windows Defender Cache Maintenance"
-    "\Microsoft\Windows\Windows Defender\Windows Defender Cleanup"
-    "\Microsoft\Windows\Windows Defender\Windows Defender Scheduled Scan"
-    "\Microsoft\Windows\Windows Defender\Windows Defender Verification"
-)
+    #"Disabling Windows Defender via Group Policies"
+    ForceRegKey "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender"
+    Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender" -Name "DisableAntiSpyware" -Type DWord -Value 1
+    Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender" -Name "DisableRoutinelyTakingAction" -Type DWord -Value 1
+    ForceRegKey "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender\Real-Time Protection"
+    Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender\Real-Time Protection" -Name "DisableRealtimeMonitoring" -Type DWord -Value 1
 
-foreach ($task in $tasks) {
-    $parts = $task.split('\')
-    $name = $parts[-1]
-    $path = $parts[0..($parts.length-2)] -join '\'
-    
-    Disable-ScheduledTask -TaskName "$name" -TaskPath "$path"
+    #"Disabling Windows Defender Services"
+    TakeownRegistry("HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\WinDefend")
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WinDefend" "Start" 4
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WinDefend" "AutorunsDisabled" 3
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WdNisSvc" "Start" 4
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WdNisSvc" "AutorunsDisabled" 3
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\Sense" "Start" 4
+    Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\Sense" "AutorunsDisabled" 3
+
+    #"Removing Windows Defender context menu item"
+    Set-Item -Path "HKLM:\SOFTWARE\Classes\CLSID\{09A47860-11B0-4DA5-AFA5-26D86198A780}\InprocServer32" -Value ""
+
+    #"Removing Windows Defender GUI / tray from autorun"
+    Remove-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Run" -Name "WindowsDefender" -ea 0
 }
-
-#"Disabling Windows Defender via Group Policies"
-ForceRegKey "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender"
-Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender" -Name "DisableAntiSpyware" -Type DWord -Value 1
-Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender" -Name "DisableRoutinelyTakingAction" -Type DWord -Value 1
-ForceRegKey "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender\Real-Time Protection"
-Set-ItemProperty -Path "HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows Defender\Real-Time Protection" -Name "DisableRealtimeMonitoring" -Type DWord -Value 1
-
-#"Disabling Windows Defender Services"
-TakeownRegistry("HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Services\WinDefend")
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WinDefend" "Start" 4
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WinDefend" "AutorunsDisabled" 3
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WdNisSvc" "Start" 4
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\WdNisSvc" "AutorunsDisabled" 3
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\Sense" "Start" 4
-Set-ItemProperty -Path "HKLM:\SYSTEM\CurrentControlSet\Services\Sense" "AutorunsDisabled" 3
-
-#"Removing Windows Defender context menu item"
-Set-Item -Path "HKLM:\SOFTWARE\Classes\CLSID\{09A47860-11B0-4DA5-AFA5-26D86198A780}\InprocServer32" -Value ""
-
-#"Removing Windows Defender GUI / tray from autorun"
-Remove-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Run" -Name "WindowsDefender" -ea 0
